requirements.txt
# 音频处理
pydub>=0.25.1
ffmpeg-python>=0.2.0

# 语音转写
openai-whisper>=20231117
torch>=2.0.0

# 翻译
transformers>=4.35.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# 本地翻译模型（可选）
# deep-translator>=1.11.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0

# Anki导出
genanki>=0.13.0

# GUI框架
PyQt5>=5.15.0
pyqt5-tools>=5.15.0

# 工具库
click>=8.1.0
tqdm>=4.65.0
pillow>=10.0.0
python-dotenv>=1.0.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0